\name{BehavioralEconomics-package}
\alias{BehavioralEconomics-package}
\alias{BehavioralEconomics}
\docType{package}
\title{
Statistical models for behavioral economics
}
\description{
Tools for model fitting, model assessment and model simulation for researchers working with behavioral economic models, especially models for intertemporal choice and decision-making under uncertainty.
}
\details{
\tabular{ll}{
Package: \tab BehavioralEconomics\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2010-12-30\cr
License: \tab Artistic-2.0\cr
LazyLoad: \tab yes\cr
}
1
}
\author{
John Myles White

Maintainer: John Myles White <jmw@johnmyleswhite.com>
}
\references{
Laibson
Prelec
}
\keyword{ package }
\examples{
1
}
